requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
function==1.2.0
idna==3.4
numpy==1.23.4
Pillow==9.3.0
requests==2.28.1
sqlparse==0.4.3
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
